Rename service package to dataset.
